some edit

+++ src/main/java/ru/bgpu/server/dto/FileInfoDto.java
@@ -4,7 +4,7 @@ import java.io.File;
 
 public class FileInfoDto {
     private String name;
-    private double size;
+    private long size;
 
     public FileInfoDto(File file) {
         this.name = file.getName();
